requirements.txt
Django
requests
django-ipware
ipython